Intermediate... do we factor out metrics to SDK?
