+++ controlplane/pkg/nsmd/nsmd_crossconnect_client.go
@@ -29,12 +29,12 @@ import (
 	"github.com/networkservicemesh/networkservicemesh/controlplane/api/registry"
 	remote "github.com/networkservicemesh/networkservicemesh/controlplane/api/remote/connection"
 	"github.com/networkservicemesh/networkservicemesh/controlplane/pkg/model"
-	"github.com/networkservicemesh/networkservicemesh/controlplane/pkg/monitor"
-	monitor_crossconnect "github.com/networkservicemesh/networkservicemesh/controlplane/pkg/monitor/crossconnect"
-	monitor_local "github.com/networkservicemesh/networkservicemesh/controlplane/pkg/monitor/local"
-	monitor_remote "github.com/networkservicemesh/networkservicemesh/controlplane/pkg/monitor/remote"
 	"github.com/networkservicemesh/networkservicemesh/controlplane/pkg/services"
 	"github.com/networkservicemesh/networkservicemesh/pkg/tools"
+	"github.com/networkservicemesh/networkservicemesh/sdk/monitor"
+	monitor_crossconnect "github.com/networkservicemesh/networkservicemesh/sdk/monitor/crossconnect"
+	monitor_local "github.com/networkservicemesh/networkservicemesh/sdk/monitor/local"
+	monitor_remote "github.com/networkservicemesh/networkservicemesh/sdk/monitor/remote"
 )
 
 const (
