+++ sdk/common/constants.go
